requirements.txt
anthropic>=0.25.0
pandas>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
tqdm>=4.65.0